Initial commit of waitless, a decent build system
This is a partial outline of the LD_PRELOAD stubs, action interface, and
data structures; see the comments for details.  I've made a lot of
effort to keep the headers and interface boundaries clean in order to
make it easier to understand.  It really is just an outline at this
point; a large fraction of the functions are NOT_IMPLEMENTED stubs.
TODOs abound.

I'm writing this in outline order (some kind of hybrid breadth / depth
first traversal of the code structure) because it's the only way I know
how to think through all the race conditions.  Waitless is essentially
an asynchronous, speculative, superscalar execution engine for
processes, so there are a lot of race conditions.

We use Skein as our cryptographic hash because it's (hopefully)
reasonably strong and significantly faster than SHA-1.  The Skein
code is a slightly cleaned subset of the code from

    http://www.skein-hash.info

128 bits should be enough for now.  The Skein code is public domain
except for some unnecessary compatibility headers which I stripped.